requirements.txt
boto3==1.4.7
botocore==1.7.41
django-storages==1.6.5
certifi==2017.11.5
chardet==3.0.4
codegen==1.0
Django==1.8
django-crispy-forms==1.7.0
django-embed-video==1.1.2
django-messages==0.5.3
django-multiselectfield==0.1.8
django-registration-redux==1.8
django-user-agents==0.3.2
djangorestframework==3.7.3
idna==2.6
mysql-connector-python==2.1.7
mysqlclient==1.3.12
numpy==1.13.3
olefile==0.44
patterns==0.3
pika==0.11.0
Pillow==4.3.0
Pygments==2.2.0
pyserial==3.4
python-memcached==1.58
PyYAML==3.12
requests==2.18.4
six==1.11.0
ua-parser==0.7.3
urllib3==1.22
user-agents==1.1.0